tehtävä 4.21

+++ part4/controllers/blogs.js
@@ -50,8 +50,24 @@ blogsRouter.post("/", async (request, response) => {
 
 // DELETE A BLOG
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({
+      error: "you do not have permission to delete this blog",
+    });
+  }
 });
 
 // UPDATE A BLOG
